There are two changes on this commit.
One is that we finalized the code to store all online availability information into a JSON string that can handle more that one link with all its related information. Notice that we need to update solrconfig.xml on the server to pick up the new field in searches.

The second change is that we are using a local data file with ProQuest data to add the links for ETDs that are available on their site. This also allows to index these records correctly as online.